@@ -36,28 +36,26 @@
   "dependencies": {
     "@nextcloud/auth": "^2.2.1",
     "@nextcloud/axios": "^2.4.0",
-    "@nextcloud/dialogs": "^5.1.2",
-    "@nextcloud/files": "^4.0.0-beta.4",
-    "@nextcloud/initial-state": "^2.1.0",
-    "@nextcloud/l10n": "^2.2.0",
+    "@nextcloud/dialogs": "^7.2.0",
+    "@nextcloud/files": "^4.0.0-beta.8",
+    "@nextcloud/initial-state": "^3.0.0",
+    "@nextcloud/l10n": "^3.4.1",
     "@nextcloud/moment": "^1.3.1",
-    "@nextcloud/password-confirmation": "^5.1.1",
+    "@nextcloud/password-confirmation": "^6.0.2",
     "@nextcloud/router": "^3.0.0",
-    "@nextcloud/vue": "^8.26.0",
+    "@nextcloud/vue": "^9.3.1",
     "@nextcloud/vue-dashboard": "^2.0.1",
-    "vue": "^2.7.16",
-    "vue-demi": "^0.14.10",
+    "vue": "^3.5.26",
     "vue-material-design-icons": "^5.3.0"
   },
   "devDependencies": {
     "@nextcloud/babel-config": "^1.0.0",
     "@nextcloud/browserslist-config": "^3.0.0",
     "@nextcloud/eslint-config": "^8.3.0",
-    "@nextcloud/stylelint-config": "^2.4.0",
-    "@nextcloud/webpack-vue-config": "^5.5.1",
-    "eslint-webpack-plugin": "^4.0.1",
+    "@nextcloud/stylelint-config": "^3.1.1",
+    "@nextcloud/webpack-vue-config": "^7.0.2",
+    "eslint-webpack-plugin": "^5.0.2",
     "stylelint-webpack-plugin": "^5.0.0",
-    "webpack": "^5.93.0",
-    "webpack-cli": "^5.1.4"
+    "webpack": "^5.93.0"
   }
 }
