@@ -11,11 +11,15 @@
  *
  */
 
-import Vue from 'vue'
-import './bootstrap.js'
+import { createApp } from 'vue'
 import AdminSettings from './components/AdminSettings.vue'
-import Tooltip from '@nextcloud/vue/dist/Directives/Tooltip.js'
-Vue.directive('tooltip', Tooltip)
 
-const View = Vue.extend(AdminSettings)
-new View().$mount('#docusign_prefs')
+import { linkTo } from '@nextcloud/router'
+import { getCSPNonce } from '@nextcloud/auth'
+
+__webpack_nonce__ = getCSPNonce() // eslint-disable-line
+__webpack_public_path__ = linkTo('integration_docusign', 'js/') // eslint-disable-line
+
+const app = createApp(AdminSettings)
+app.mixin({ methods: { t, n } })
+app.mount('#docusign_prefs')
