@@ -70,7 +70,7 @@
 import CloseIcon from 'vue-material-design-icons/Close.vue'
 import DocusignIcon from './icons/DocusignIcon.vue'
 
-import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
+import NcButton from '@nextcloud/vue/components/NcButton'
 
 import { loadState } from '@nextcloud/initial-state'
 import { generateUrl } from '@nextcloud/router'
